+++ Octans.Core/Downloads/DomainBandwidth.cs
