+++ Octans.Core/Downloads/DownloadRequest.cs
@@ -1,0 +1,12 @@
+namespace Octans.Core.Downloaders;
+
+public class DownloadRequest
+{
+    public required string Url { get; set; }
+    public required string DestinationPath { get; set; }
+    
+    /// <summary>
+    /// Higher numbers = higher priority
+    /// </summary>
+    public int Priority { get; set; }
+}
