+++ Octans.Core/Downloads/DownloadService.cs
