+++ Octans.Data/Models/DownloadState.cs
@@ -1,0 +1,12 @@
+namespace Octans.Core.Downloaders;
+
+public enum DownloadState
+{
+    Queued,
+    WaitingForBandwidth,
+    InProgress,
+    Paused,
+    Completed,
+    Failed,
+    Canceled
+}
